WIP: Added a 'status' view.
Both the celery-process and potential manifest will share the same UUID. This commit adds the following discussed functionality:

POST-ing to /manifests/ starts a celery process and returns a 202 and link to the status page.
If status=SUCCESS then /status/[id] returns a 303 and a Location link to the manifest's page.
Else returns a dump of the celery process info: Status and Traceback and maybe other stuff in the future.