Fixed: URLs will only accept 36-char version of UUID.
That's the UUID with the hyphens in it. I had it accepting 32-36 long strings as UUIDs, but seeing as all the uuids getting passed around are being passed around by computers, having the 'option' to not include the hypens is not that helpful.

Pythons UUID representation has hypens, so thats that.