WIP: Switched to class-based manifest instantiation.
I was writing a ton of interdependent functions that kept passing dicts between each other, and I suddenly remembered what classes are. This should make it much simpler to deal with the multiple sequential steps that need to be taken to validate, index, and create a db entry for manifests.