WIP: Reworked metadata system.
We now have a hash table of metadata synonyms being stored in settings.py. Indexer will search check for a match, but if one isnt found, will dump the value of the data in a multivalued metadata field so it can still be used for keyword searching.

Commented out online validator, as it was returning erroneous results.