WIP: Normalized status responses.
Put some status codes in settings so we can have some consistency all along the backend.

Now the following status codes are returnes:

Http404: The job is 'PENDING', i.e. celery is not aware of it.

status = -1: The job resulted in an error. 'trace' (if it was an uncaught exception) or 'error' (if it was an indexing/validation error) will be included in this case.

status = 0: The job has been recieved and is unfinished. This is 'in progress'.

status = 1: The job was succesful. This comes with an http303 and a 'location' in the header and body to see the detail