WIP: Improved WIPManifest class.
WIPManifest has had most of its methods made private, and is given a 'create' method, which will ensure the various steps (getting, validating, indexing, db-stuff) are run in the correct order.

WIPManifest now checks for existing DB entries with the same remote_url as itself. The link to this existing DB entry will be returned to the frontend, rather than deleting it and creating a new page. Also checks for and deletes duplicate DB entries at this stage.

WIPManifest now checks for duplicate documents in solr based on remote_url and deletes any it finds before indexing itself.