Fixed: Removed schema.xml
Apparently if I want solr to manage the schema dynamically, I'm not supposed to have a schema.xml. Instead, this managed schema should be the only thing around, which can be modified in (I think) the same manual way, or through the schema API - which is actually kind of sweet.