WIP: Continued improvements to WIPManifest.
Modified error-prone methods to raise a ManifestImportError exception as soon as they encounter an error.

Placed above calls in a try block in create() which will catch ManifestImportError and return False to the caller. Trying to make it so the create() method returns True if everything worked and False if an error was hit - maybe this is too much like C.

Added a __parse_remote_url() function which will check that the provided url has a scheme and is properly formatted. I'm thinking a parsed url (only netlocation + path) might be a good thing to save to do fuzzy lookups for dupe detection.