Fixed: Ensure Diva is destroyed on unmount
Previously, the code which uninitialized Diva was chained to the
promise for the loading of the Diva module. This was unnecessary
because it's possible to verify synchronously if the Diva instance
is initialized. Moreover, by deferring Diva's destroy call it
introduced the possibility that the component could be unmounted
without Diva ever being destroyed. This led to errors when Diva
was reinitialized since the previous Diva events were never unbound.

Fixes #16