Refactor: Use ES7 static class properties
Previously, class attributes were being declared using an
ES5 getter combined with an ES2016 decorator. This was a bit verbose
and awkward, and used a feature (decorators) which is still only
a proposal. This commit uses another proposed feature, static
class properties, instead. This should be no harder to maintain
and should be clearer.