New: Move jspm-server out of npm devDependencies
Previously, jspm-server was a development dependency for Misirlou,
meaning it would always be installed. However, it's a fairly
heavyweight dependency that isn't needed to just build the JS
files. This commit moves its installation into a separate npm
script, which is awkward but works as a way to make it an optional
dependency.