New: Support locally configurable Django settings
Add support for a local_settings module which overrides the
defaults and is not checked in to version control. While we may
move to a different configuration strategy in the future, this
is an easy way to allow local configration and to keep security-
sensitive values out of the version control.

An example local settings file has been checked in as
misirlou/example-local_settings.py (this is by design a name which
is hard to import).