New: Added test solr module.
This module will be used during unit testing to ensure that the current solr settings are good.
This core's settings are symlinked to the main cores settings, so it should provide an isolated
core to test on, while being completely up to date in terms of configuration with the main core.