WIP: Solr indexing
Worked on the solr indexing algorithm a bit to solve some errors and hopefully improve its behaviour. No longer throws errors when confrontred with a list for labels. As sone languages are not recognized by solr, I've currently limited the languages which solr will make new fields for to english, french, italien and german.