WIP: Implemented collection importing.
Apologies for the massive commit, but implementing this revealed some assumptions in the metadata indexing loop that were too general, and thus were causing errors. It seems that every combination under the sun of label/value list/dict/str combinations exists, even ones that seemed to make no sense to me (Having a list of labels, all in the same language, specifying a single value - or having a list of values for fields like 'title' which are not specifying multiple languages, just multiple titles).

Wrote an Importer class which figures out if the given url is a single manifest or collection. If it is a collection, it recursively collects all manifest urls from the collection and sub collection, then returns this list to be imported. I chose to keep collection importing on a single thread, as even modest collections can have >500 manifests, which would completely stop any single manifest jobs from getting a chance on a celery thread for a few minutes.

The status json objects will reflect the fact that a collection is being imported as soon as the list of manifests is finished being compiled. Along with the regular status=1, the dict will contain a 'progress' key which contains numbers for the current and total manifests processed.