Fixed: Added proper status for collection importing.
The succesful import status now contains a 'type' field, which specified if a manifest or collection was imported. If a collection was imported, the status code is 200 instead of see other (as there is not an obvious manifest to redirect to). This status will contain the total number of manifests on which import was attempted [total], the number which succeeded [succeeded], a list of URL's to the manifests that failed [failed].

There is also a large dict called [trace] (accepting suggestions for better names). The key's in trace are the remote urls of the manifests. Each entry has the warnings/errors that effected that manifest, and a [location] link if the manifest imported succesfully.