Fixed: spellchecker stopwords.
The field that spell-checking and suggestions are sourced from will check against a giant list of stopwords for all the languages we index. This will fix the error where 're c en si on' is a suggestion for the word 'recension', since the extremely common 'en' and 'si' are no longer considered.

Fixes: #28