New: Comparing netlocs of urls.
If the remote_url does not have the same netloc as the @id in the document,
then we will continue to use the remote_url rather than the @id during the
import process. This is to stop a situations where a faulty manifest F
could replace a genuine manifest G just by copying G's url into it's @id
field.