WIP: Improved concurrent imports.
It turns out the best thing was to determine if I need to do a large concurrent import on the main thread before sending the tasks to celery. This could be further improved by putting the initial http request in its own thread, so if it times out it doesn't lock up a django thread for a minute.

Also made a lot of sense to move solr commits onto a celery beat schedule. I wish I had investigated that solution sooner.

It appears I am much closer to my goals for this process with this commit. A lot of code I wrote has been cleaned up, as I have pursued different ways to solve this issue.