WIP: Updated to newset restframework.
This caused an infinite recursion bug in 'view_data' key which is being injected into the context. For now it is fixed by copying the 'data' object into it's value, rather than simple referencing it, but I do not know if this is the best solution. @wabain should definitly have a look!