New: Changed response of /manifests/recent (IMPORTANT)
More API breaking changes for @wabain. Instead of querying over the DB on the
non-indexed timestamp, we get recent manifests from solr. This should be faster,
and has the advantage of including the indexed thumbnail in the response. This
will save you from having to fetch every manifest on the front page just to
display a thumbnail.

The response will be identical to a query on the main page, it is just hadcoded
to sort by descending created_timestamp (you can't request this over http with
the main page search right now... It only looks for a query).