Fixed: Automatic thumbnail setting.
The thumbnail setting function will now delete an 'item' key under the
resource it finds, if such a key exists. Since the point of this fn
is just to grab the first image it can find to display as a thumbnail,
having multiple representations of this image is unnesesary. Only the
top level is needed.