Refactor: Normalize manuscript summary objects
Use the same object shape and props checks for the manifest summary
objects in the manifest cascade on the landing page and the
search results. Previously the search results were placed in an
Immutable.js record, which is overkill because they're never going
to be updated. Just freeze them instead to maintain the immutability
guarantee.