WIP: Put manifest schema inside of class.
This will make it easier to run the validator in different modes...
e.g., a 'strict' mode which catches all errors, and a 'flexible' mode
which corrects minor errors as it validates. It will also make it possible
to collect warnings that are not show-stoppers, as the validator can access
the state of the object it is in (instead of being top-level in a module).