WIP: Began work on manifest tester.
This tester is intended to be used both automatically and manually
to iterate over the manifests in our database and report issues on them.

The most important thing ia a hash comparison, to check that the file we
have indexed still represents the file hosted by the library, but this
tester will also check for numerous other errors and warnings. It is also
supposed to be configurable, so that in the future we can run specific
tests on the database relatively easily.