Fixed: Teseter handles thumbnails better.
ValueError is raised when a string can't be parsed to json - so this means that there is some error
that would completely prevent the thumbnail from being displayed.

TypeError is raised when trying a __getitem__ on a string, which may be the case if the thumbnail
is a single uri instead of an object.