WIP: Fixed lack of @context disqualifying manifest.
If a manifest fails a check for IIIF image service spec compliance
in its default sequence, it will now raise a warning by default in
the tester.

The @context key will no longer be required in the 'flexible'
validator mode. We are assuming the context is a IIIF image service.