WIP: Added outline for handling individual librarires.
Since it seems like bad manifests will probably always be an issue,
it would help to have an orderly place to specify how we handle each
library's quirks.

I think in the long run this will be better than continually expanding
the 'flexible' validation. It seems to me that by continuing to relax
our validator to handle libraries with bad manifests, it will eventually
degrade to a point where it is an imcomprehensible mess of corner cases
and exceptions.

Instead, we keep the validator generic and strict, but introduce library
specific exceptions to be patched in at runtime. This will make it much
easier to reason about the validator.