WIP: Added some settings for importing liber.
This is not a very good command. It does not make sense to make a very good one
though, if we still need to develop a clear a workflow for going from manifest to
mei to csv to solr. Since all the documents we have are somewhat unique (for example,
with respect to page numbering), it makes things difficult.