Temp fix: Removing diva toolbar.
Diva does not clean up its toolbar when you destroy it. And, it looks like
there is no destroy method for the toolbar in the source. This will just
clear out the toolbar before a new document is loaded.

Probably proper cleanup by diva would be preferable in the long run!

References #78