New: IIIFValidator class.
Pulled validation intro point out from ManifestValidator
and replaced it with funcitonality of IIIFValidator class.

The purpose of this is to allow Manifests to be part of a
greater structre. The 'top-level' validator which all the
other validators refer to should not actually be a validator
for a specific field, but instead should be a more abstract
class that manipulates the actual validators.

This way, eventually we could extend out to do validation
of structures which contain manifests. It's also clearer, as
the ManifestValidator is no longer more 'special' than the
CanvasValidator or SequenceValidator. It's just another in the
hierarchy of structures that are validated.