New: Page min-width and media queries for input fields
Bootstrap has been removed from the navbar in order to put a min-width
on the root element. Also, the input fields in the navbar now have 3
different sizes which vary based on the width of the viewport (done
using media queries). Bugs related to small layouts have also been fixed.

Refs #77, Closes #58