New: Added generalizable search scraper.
Created a class to do general search scraping. It contains
two methods which will be over-riden on a per-site basis in
order to scrape IIIF manifests out of search results.

The scrape method will handle rate limiting and create
multiple threads to handle the scraping. Class is flexible
enough that a single scraper can produce child scrapers and
share a rate limiter with them (if a searh result is a link
to another search).

Updated the gallica scraper to use this new class.