Fixed: Highlights from spellcheck_txt
If the only highlight avaliable on a search is
from the spellcheck_txt field, it means that the match
is on some unkown metadata field. In this case, the
field will be renamed to 'Metadata' in the results.

If there are other fields with highlights, the
highlight on spellcheck_txt will be ignored.