New: Added phrase suggestions.
A new phrase suggestion suggester has been added, which will
try to match what you've typed as a prefix to the title or label
or a manifest. These suggestions have been seamlessly combined
with the existing 'last-word' based spell checking.

The logic to strip the 'last-word' out of a phrase has been
moved into a regex run by solr. This also means we can query
both suggesters in one request.