New: Added signals for succeeded import.
Added a signal and handler for when a manifest is imported
(note, this is crucially different from when a manifest is
saved).

This signal will only fire when an import is entirely
successful, meaning that manifests which are re-imported will
only fire the signal if they are updated, and new manifests
will only fire the signal if they are fully ready to go.

All that is needed to integrate this in rodan is to fill out
the handler in /models/manifest.py with whatever behaviour
is needed.