New: Added distributed respectful requester.
Basically this requester maintains a distributed set of locks
and timestamps for every domain it gets a request to get. The idea
is that any number of celery processes can be working on importing stuff
from any set of domains, and we can ensure that we never hit each domain more
than x times per second.

The requester first looks for a Crawl-delay time in the domain's robot.txt,
then saves this for future use. If no Crawl-delay is defined, we default to
1 second between requests.

Implemented using Redis, which we are already using as our celery
backend, so really no new requirements.

This solves the problem of us slamming servers whenever we want to reimport
or test a bunch of manifests.

There remains a problem that we have no way to re-assign workers to tasks.
For instance, if domain x has a crawl delay of 30 seconds, if a worker
gets a request to this domain and the last request went through 5 seconds
ago, the worker has no way to put the task back in the pool and ask for a
new one (there may be other requests to other domains pending). The worker
is doomed to wait the 25 seconds.

This is inefficient, but I think probably better than slamming servers.