New: Command to partition last_tested time.
Since I haven't run tests on any of the manifests in a year,
and running a test on one manifest involves about 5 get requests
to a server, this commad is designed to partition
the last_tested dates of the manifests over the last 14 days.

Then we can run a command to test any manifests that haven't been
tested in more than 14 days. If we do this every day, it evenly
spreads out the testing.