Fixed: Fixed small issues on about page.
The about page is not actually connected to the one-page app, so
it's just emulating the Navbar at the top of the page. This is
not great, because we have two navbar.js files, one for the main
project and one for the about page.

This is the state things are currently in and I don't have to time
to stitch things together. So, in the meanwhile, I've at least cleaned
up some visual issues on the page to make it more presentable.